+++ app/components/omnibox/controllers/point-controller.js
@@ -20,13 +20,13 @@ angular.module('lizard-nxt')
   'ClickFeedbackService',
   'UtilService',
   'MapService',
+  'DataService',
   'State',
 
-  function ($scope, $q, LeafletService, TimeseriesService, ClickFeedbackService, UtilService, MapService, State) {
+  function ($scope, $q, LeafletService, TimeseriesService, ClickFeedbackService, UtilService, MapService, DataService, State) {
 
     var GRAPH_WIDTH = 600;
     $scope.box.content = {};
-    $scope.box.showFullTable = false;
 
     /**
      * @function
@@ -37,7 +37,7 @@ angular.module('lizard-nxt')
       ClickFeedbackService.drawCircle(MapService, here);
       ClickFeedbackService.startVibration(MapService);
       var aggWindow = State.temporal.aggWindow;
-      var promise = $scope.fillBox('point', {
+      var promise = $scope.fillBox({
         geom: here,
         start: State.temporal.start,
         end: State.temporal.end,
@@ -216,6 +216,8 @@ angular.module('lizard-nxt')
 
     // Clean up stuff when controller is destroyed
     $scope.$on('$destroy', function () {
+      DataService.reject();
+      $scope.box.content = {};
       ClickFeedbackService.emptyClickLayer(MapService);
     });
   }