+++ app/lib/util-service.js
@@ -643,6 +643,15 @@ angular.module('lizard-nxt')
     }
   };
 
+  this.TIMELINE_LEFT_MARGIN = 60;
+  this.TIMELINE_RIGHT_MARGIN = 40;
+
+  this.getCurrentWidth = function () {
+    return window.innerWidth - (
+      this.TIMELINE_LEFT_MARGIN + this.TIMELINE_RIGHT_MARGIN
+    );
+  };
+
 
   // Add comparator to sort lists on multiple properties to D3.
   (function () {