+++ app/components/data-menu/services/layer-service.js
@@ -242,6 +242,10 @@ angular.module('lizard-nxt')
           };
 
           options = options || {};
+          
+          // Pass layer options to the services making the request.
+          // RasterServices uses this to add options.styles.
+          angular.extend(options, this.options);
           options.agg = this.aggregationType;
 
           return wantedService.getData(this, options)