+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog of lizard-nxt client
 Unreleased (1.2.dev) (XXXX-XX-XX)
 ---------------------------------
 
+- timeline: it's "netto width" is made available through UtilService
+
+- timeline: zoom buttons working
+
 - Omnibox rain graph syncs x-axis to timeline.
 
 - Grunt release script.