+++ CHANGES.rst
@@ -4,9 +4,9 @@ Changelog of lizard-nxt client
 Unreleased (1.2.dev) (XXXX-XX-XX)
 ---------------------------------
 
-- timeline: it's "netto width" is made available through UtilService
+- timeline: it's "netto width" is made available through UtilService.
 
-- timeline: zoom buttons working
+- timeline: zoom buttons working.
 
 - Omnibox rain graph syncs x-axis to timeline.